site: update
Little did you all know, this was more of a documentation effort for
future Dilyn than a instruction manual for new people type of thing.

I hope you found the series useful, but this is all mostly just as a
saving point for quick and easy redeploymenet if I happen to fsck my
server up something fierce.